+++ examples/tree_2d_fdsbp/elixir_euler_convergence.jl
@@ -6,51 +6,57 @@ using Trixi
 
 ###############################################################################
 # semidiscretization of the linear advection equation
-
-advection_velocity = (0.2, -0.7)
-equations = LinearScalarAdvectionEquation2D(advection_velocity)
+equations = CompressibleEulerEquations2D(1.4)
 
 initial_condition = initial_condition_convergence_test
+source_terms = source_terms_convergence_test
 
-D_SBP = derivative_operator(SummationByPartsOperators.MattssonNordström2004(),
-                            derivative_order=1, accuracy_order=4,
-                            xmin=0.0, xmax=1.0, N=100)
-solver = FDSBP(D_SBP,
-               surface_integral=SurfaceIntegralStrongForm(flux_lax_friedrichs),
-               volume_integral=VolumeIntegralStrongForm())
+D_upw = upwind_operators(SummationByPartsOperators.Mattsson2017,
+                         derivative_order=1,
+                         accuracy_order=4,
+                         xmin=-1.0, xmax=1.0,
+                         N=16)
+flux_splitting = splitting_steger_warming
+solver = FDSBP(D_upw,
+               surface_integral=SurfaceIntegralUpwind(flux_splitting),
+               volume_integral=VolumeIntegralUpwind(flux_splitting))
 
 coordinates_min = (-1.0, -1.0)
 coordinates_max = ( 1.0,  1.0)
 mesh = TreeMesh(coordinates_min, coordinates_max,
-                initial_refinement_level=1,
+                initial_refinement_level=3,
                 n_cells_max=30_000,
                 periodicity=true)
 
-semi = SemidiscretizationHyperbolic(mesh, equations, initial_condition, solver)
-
+semi = SemidiscretizationHyperbolic(mesh, equations, initial_condition, solver, source_terms = source_terms_convergence_test)
 
 ###############################################################################
 # ODE solvers, callbacks etc.
 
-tspan = (0.0, 1.0)
+tspan = (0.0, 2.0)
 ode = semidiscretize(semi, tspan);
 
 summary_callback = SummaryCallback()
 
-analysis_interval = 100
+analysis_interval = 1000
 analysis_callback = AnalysisCallback(semi, interval=analysis_interval,
                                      extra_analysis_integrals=(energy_total,))
 
 alive_callback = AliveCallback(analysis_interval=analysis_interval)
 
+save_solution = SaveSolutionCallback(interval=1000,
+                                     save_initial_solution=true,
+                                     save_final_solution=true,
+                                     solution_variables=cons2prim)
+
 callbacks = CallbackSet(summary_callback,
                         analysis_callback,
+                        save_solution,
                         alive_callback)
-
 
 ###############################################################################
 # run the simulation
 
-sol = solve(ode, RDPK3SpFSAL49(), abstol=1.0e-9, reltol=1.0e-9,
+sol = solve(ode, SSPRK43(), abstol=1.0e-9, reltol=1.0e-9,
             save_everystep=false, callback=callbacks)
 summary_callback()