+++ src/auxiliary/math.jl
@@ -192,5 +192,25 @@ julia> min(2, 5, 1)
 
 
 
+"""
+    positive_part(x)
+
+Return `x` if `x` is positive, else zero. In other words, return
+`(x + abs(x)) / 2` for real numbers `x`.
+"""
+@inline function positive_part(x)
+  return max(x, zero(x))
+end
+
+"""
+    negative_part(x)
+
+Return `x` if `x` is negative, else zero. In other words, return
+`(x - abs(x)) / 2` for real numbers `x`.
+"""
+@inline function negative_part(x)
+  return min(x, zero(x))
+end
+
 
 end # @muladd